Add files via upload
Book recommendation system finds top 10 most similar books to what you search for. This is a website for book recommendation using TF-IDF vectorisation. Starting from the dataset first all the core details like title, subtitle, author, description are added into new column in the form of a list. The list is cleaned with the help of port stemmer and other python commands and on which tf-idf score is calculated. Then with the help of cosine similarity each book is compared with all the other books from the list and top 10 similar books are fetched from the list and shown with the most similar first